+++ source/libs/data/objects/components/collisions/BoxCollider.h
@@ -49,7 +49,7 @@ public:
 private:
   std::array<glm::vec3, boxVertices.size()> m_transformedBoxVertices{};
 
-  uint8_t m_currentTransformUpdateID = 255;
+  uint64_t m_currentTransformUpdateID = 255;
 
   // Editing the collider's own offset doesn't bump the transform's update id, so force a mesh rebuild.
   bool m_meshDirty = true;
